+++ changelog.md
@@ -45,5 +45,10 @@
 - Finalized the ki predictor on the regression only pipeline.
 - Saved the models created for inference.
 
-## 3/30/2013 Changes
+## 3/30/2023 Changes
 - Changed the variance on the regression only model to 85
+
+## 4/3/2023 Changes
+- Added graphing fucntion to regression_only.py that graphs the all of the resutls from hyper-
+    parameter tuning on the best model that we chose.  No changes were made to the final selected
+    model